+++ src/Formatter.ts
@@ -20,6 +20,10 @@ export class Formatter {
             includeWhitespace: true
         });
 
+        if (options.formatMultiLineObjectsAndArrays) {
+            tokens = this.formatMultiLineObjectsAndArrays(tokens);
+        }
+
         if (options.compositeKeywords) {
             tokens = this.formatCompositeKeywords(tokens, options);
         }
@@ -49,6 +53,116 @@ export class Formatter {
         return outputText;
     }
 
+    /**
+     * Determines if the current index is the start of a single-line array or AA.
+     * Walks forward until we find the equal number of open and close curlies/squares, or a newline
+     */
+    private isStartofSingleLineArrayOrAA(tokens: Token[], currentIndex: number, openKind: TokenKind, closeKind: TokenKind) {
+        let openCount = 0;
+        for (let i = currentIndex; i < tokens.length; i++) {
+            let token = tokens[i];
+            if (token.kind === openKind) {
+                openCount++;
+            } else if (token.kind === closeKind) {
+                openCount--;
+            }
+            if (openCount === 0) {
+                return true;
+            } else if (token.kind === TokenKind.Newline) {
+                return false;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Find the matching closing token for open square or open curly
+     */
+    private getCloser(tokens: Token[], currentIndex: number, openKind: TokenKind, closeKind: TokenKind) {
+        let openCount = 0;
+        for (let i = currentIndex; i < tokens.length; i++) {
+            let token = tokens[i];
+            if (token.kind === openKind) {
+                openCount++;
+            } else if (token.kind === closeKind) {
+                openCount--;
+            }
+            if (openCount === 0) {
+                return token;
+            }
+        }
+    }
+
+    private isMatchingDoubleArrayOrArrayCurly(tokens: Token[], currentIndex: number) {
+        let token = tokens[currentIndex];
+        let nextNonWhitespaceToken = this.getNextNonWhitespaceToken(tokens, currentIndex, true);
+        //don't separate multiple open/close pairs
+        if (
+            //is open array
+            token.kind === TokenKind.LeftSquareBracket &&
+            //there is another token on this line
+            nextNonWhitespaceToken &&
+            //is next token an open array or open object
+            (nextNonWhitespaceToken.kind === TokenKind.LeftSquareBracket || nextNonWhitespaceToken.kind === TokenKind.LeftCurlyBrace)
+        ) {
+            let closingToken = this.getCloser(tokens, currentIndex, TokenKind.LeftSquareBracket, TokenKind.RightSquareBracket);
+            //look at the previous token
+            let previous = this.getPreviousNonWhitespaceToken(tokens, tokens.indexOf(closingToken!), true);
+            if (previous && (previous.kind === TokenKind.RightSquareBracket || previous.kind === TokenKind.RightCurlyBrace)) {
+                return true;
+            }
+        }
+    }
+
+    /**
+     * Standardize multi-line objects and arrays by inserting newlines after leading and before trailing.
+     */
+    private formatMultiLineObjectsAndArrays(tokens: Token[]) {
+        for (let i = 0; i < tokens.length; i++) {
+            let token = tokens[i];
+            let openKind: TokenKind | undefined;
+            let closeKind: TokenKind | undefined;
+
+            if (token.kind === TokenKind.LeftCurlyBrace) {
+                openKind = TokenKind.LeftCurlyBrace;
+                closeKind = TokenKind.RightCurlyBrace;
+            } else if (token.kind === TokenKind.LeftSquareBracket) {
+                openKind = TokenKind.LeftSquareBracket;
+                closeKind = TokenKind.RightSquareBracket;
+            }
+
+            //move contents to new line if this is a multi-line array or AA
+            if (
+                //is open curly or open square
+                openKind && closeKind &&
+                //is a multi-line array or AA
+                !this.isStartofSingleLineArrayOrAA(tokens, i, openKind, closeKind) &&
+                //there is extra stuff on this line
+                this.getNextNonWhitespaceToken(tokens, i, true) &&
+                //is NOT array like `[[ ...\n ]]`, or `[{ ...\n }]`)
+                !this.isMatchingDoubleArrayOrArrayCurly(tokens, i)
+            ) {
+                tokens.splice(i + 1, 0, <any>{
+                    kind: TokenKind.Newline,
+                    text: '\n'
+                });
+                let closingToken = this.getCloser(tokens, i, openKind, closeKind);
+                let closingTokenKindex = tokens.indexOf(closingToken!);
+
+                i++;
+
+                //if there's stuff before the closer, move it to a newline
+                if (this.getPreviousNonWhitespaceToken(tokens, closingTokenKindex, true)) {
+                    tokens.splice(closingTokenKindex, 0, <any>{
+                        kind: TokenKind.Newline,
+                        text: '\n'
+                    });
+                }
+            }
+        }
+        return tokens;
+    }
+
     private dedupeWhitespace(tokens: Token[]) {
         for (let i = 0; i < tokens.length; i++) {
             let currentToken = tokens[i];
@@ -74,7 +188,7 @@ export class Formatter {
                 //is this a composite token
                 CompositeKeywords.includes(token.kind) &&
                 //is not being used as a key in an AA literal
-                nextNonWhitespaceToken && nextNonWhitespaceToken.kind !== TokenKind.Colon &&
+                (!nextNonWhitespaceToken || nextNonWhitespaceToken.kind !== TokenKind.Colon) &&
                 //is not being used as an object key
                 previousNonWhitespaceToken?.kind !== TokenKind.Dot
             ) {
@@ -222,7 +336,7 @@ export class Formatter {
                         //if this is an indentor token
                         IndentSpacerTokenKinds.includes(token.kind) &&
                         //is not being used as a key in an AA literal
-                        nextNonWhitespaceToken.kind !== TokenKind.Colon
+                        nextNonWhitespaceToken && nextNonWhitespaceToken.kind !== TokenKind.Colon
                     ) {
                         //skip indent for 'function'|'sub' used as type (preceeded by `as` keyword)
                         if (
@@ -239,8 +353,8 @@ export class Formatter {
                         if (
                             //if this is an open square
                             token.kind === TokenKind.LeftSquareBracket &&
-                            //the next token is an open curly
-                            nextNonWhitespaceToken.kind === TokenKind.LeftCurlyBrace &&
+                            //the next token is an open curly or open square
+                            (nextNonWhitespaceToken.kind === TokenKind.LeftCurlyBrace || nextNonWhitespaceToken.kind === TokenKind.LeftSquareBracket) &&
                             //both tokens are on the same line
                             token.range.start.line === nextNonWhitespaceToken.range.start.line
                         ) {
@@ -251,7 +365,7 @@ export class Formatter {
                         //this is an outdentor token
                         OutdentSpacerTokenKinds.includes(token.kind) &&
                         //is not being used as a key in an AA literal
-                        nextNonWhitespaceToken.kind !== TokenKind.Colon &&
+                        nextNonWhitespaceToken && nextNonWhitespaceToken.kind !== TokenKind.Colon &&
                         //is not a method call
                         !(
                             //certain symbols may appear next to an open paren, so exclude those
@@ -492,7 +606,7 @@ export class Formatter {
     ) {
         let i = 0;
         let token: Token = undefined as any;
-        let nextNonWhitespaceToken: Token = undefined as any;
+        let nextNonWhitespaceToken: Token | undefined;
         const setIndex = (newValue) => {
             i = newValue;
             token = tokens[i];
@@ -507,14 +621,14 @@ export class Formatter {
             {
                 let parenToken: Token | undefined;
                 //look for anonymous functions
-                if (token.kind === TokenKind.Function && nextNonWhitespaceToken.kind === TokenKind.LeftParen) {
+                if (token.kind === TokenKind.Function && nextNonWhitespaceToken?.kind === TokenKind.LeftParen) {
                     parenToken = nextNonWhitespaceToken;
 
                     //look for named functions
-                } else if (token.kind === TokenKind.Function && nextNonWhitespaceToken.kind === TokenKind.Identifier) {
+                } else if (token.kind === TokenKind.Function && nextNonWhitespaceToken?.kind === TokenKind.Identifier) {
                     //get the next non-Whitespace token, which SHOULD be the paren
                     let parenCandidate = this.getNextNonWhitespaceToken(tokens, tokens.indexOf(nextNonWhitespaceToken));
-                    if (parenCandidate.kind === TokenKind.LeftParen) {
+                    if (parenCandidate?.kind === TokenKind.LeftParen) {
                         parenToken = parenCandidate;
                     }
                 }
@@ -602,7 +716,7 @@ export class Formatter {
             }
 
             //empty parenthesis (user doesn't have this option, we will always do this one)
-            if (token.kind === TokenKind.LeftParen && nextNonWhitespaceToken.kind === TokenKind.RightParen) {
+            if (token.kind === TokenKind.LeftParen && nextNonWhitespaceToken?.kind === TokenKind.RightParen) {
                 this.removeWhitespaceTokensBackwards(tokens, tokens.indexOf(nextNonWhitespaceToken));
                 //next loop iteration should be after the closing paren
                 setIndex(
@@ -662,9 +776,6 @@ export class Formatter {
     /**
      * Get the first token after the index that is NOT Whitespace
      */
-    private getNextNonWhitespaceToken(tokens: Token[], index: number, stopAtNewLine: true): Token | undefined;
-    private getNextNonWhitespaceToken(tokens: Token[], index: number, stopAtNewLine: false): Token;
-    private getNextNonWhitespaceToken(tokens: Token[], index: number): Token;
     private getNextNonWhitespaceToken(tokens: Token[], index: number, stopAtNewLine = false) {
         for (index += 1; index < tokens.length; index++) {
             let token = tokens[index];
@@ -675,8 +786,6 @@ export class Formatter {
                 return tokens[index];
             }
         }
-        //if we got here, we ran out of tokens. Return the EOF token
-        return tokens[tokens.length - 1];
     }
 
     /**